# MINOR: avoid unnecessary gettimeofday in ReplicaFetcher