This PR adds a Replica Fetcher benchmark for use in improving the performance and CPU usage of the ReplicaFetcher. In this initial PR I chose to only address one minor performance issue - unnecessary gettimeofday calls in steady state (when partitions are not delayed). I realize there are ways to improve the performance of gettimeofday calls, but it does require action on the part of the user, so I believe this optimization is still worthwhile. I have further performance optimizations that I will address in future PRs once the benchmark has been added.

Benchmarked on r5.xlarge:
```
Before:
Benchmark                                  (partitionCount)  Mode  Cnt        Score       Error  Units
ReplicaFetcherThreadBenchmark.testFetcher               100  avgt   15    28406.454 ±   614.140  ns/op
ReplicaFetcherThreadBenchmark.testFetcher               500  avgt   15   144999.326 ±  6774.986  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              1000  avgt   15   316551.913 ±  3331.297  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              5000  avgt   15  2507985.580 ± 12476.098  ns/op

After:
Benchmark                                  (partitionCount)  Mode  Cnt        Score       Error  Units
ReplicaFetcherThreadBenchmark.testFetcher               100  avgt   15    21877.395 ±   670.745  ns/op
ReplicaFetcherThreadBenchmark.testFetcher               500  avgt   15   123234.307 ±  4083.449  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              1000  avgt   15   297592.869 ±  2629.679  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              5000  avgt   15  2258424.746 ± 10646.843  ns/op
```