@ijuma 2019-09-30
Is the improvement consistent? r5s have fast gettimeofday calls and that's why it seems like the improvement is quite small in many cases.