@lbradstreet 2019-10-01
@ijuma yes, the improvement is consistent. I'm going to close this PR in favor of a more substantial PR with additional performance improvements.